finish 2

+++ index.php
@@ -49,7 +49,7 @@ $tasks = $statement->fetchAll(PDO::FETCH_ASSOC);
                     <td>
                         <a href="show.php?id=<?= $task['id'];?>" class="btn btn-info">Show</a>
                         <a href="edit.php?id=<?= $task['id'];?>" class="btn btn-warning">Edit</a>
-                        <a href="delete.php?id=<?= $task['id'];?>" class="btn btn-danger">Delete</a>
+                        <a onclick="return confirm('Are you sure?')" href="delete.php?id=<?= $task['id'];?>" class="btn btn-danger">Delete</a>
                     </td>
                 </tr>
                 <?php endforeach;?>
